[research] LibraChain
This first commit contains a proof of the safety of the LibraBFT consensus, namely the "no fork" property that says that among two blocks considered committed, one must be a parent of the other.

The property's proof is in Properties/Safety.v, and the Structures directory contains all the necessary definitions.